+++ RELEASENOTES.md
@@ -1,3 +1,14 @@
+## [0.5.0]
+
+## Added
+
+- The ability to set the settings from the request
+
+## Fixed
+
+- If there is a title at small panel sizes, the value is out of view
+- If the value is greater than the range, the color of the value is the same as the color of the last range
+
 ## [0.4.1]
 
 ### Fixed